removed usage of shell_exec
large parts of the system relied on the usage of shell_exec. This has now been ported to openssl-wrapper in php.

A change in how a CSR is inspected must be noted as PHP is unable to import a foreign CSR to a resource which the CSR-export function relies upon.
As a workaround, the system shows the fields from the subject.

Signed-off-by: Henrik Austad <henrik.austad@gmail.com>