create_keyscript should use server_loc based on actual server-loc, not a preconfigured location
When creating a new script, the create_keyscript got it's value from the confusa_config. This worked ok when we had only one vhost.
once we started running multiple vhosts on a single install, this proved to be wrong. Instead, the script should figure out the server_loc based on $_SERVER values and adapt.

Signed-off-by: Henrik Austad <henrik.austad@gmail.com>