Use custom fields to find ca-cert
Allow config to specify where the ca-cert is stored, and the name for
it. Also, use config to determine O and OU

The old function usage is not needed, only use cc_help (old help, as
help is the swithc to trigger help for create_cert.sh, hence cc_help).

Signed-off-by: Henrik Austad <henrik@austad.us>