Move to new auth_url scheme
When a client uploads a new CSR, it is tied to a corresponding auth_url.
Previously, this was a random string, but has now been changed to be based
on the public key.

This is done by taking the sha1sum of the public-key (in base64-ascii form),
and then the individual confusa-instance can choose the length of the substring.

The full length (40 characters) will be the preferred choice, but can
be confusing to EE's. Thus, it is left as a config-parameter.

Signed-off-by: Henrik Austad <henrik@austad.us>