When constructing the CommonName for the DN, use a valid CN
The profile for SLCS and MICS clearly state that the CN must be a name
that provide meaningful information. As this does not necessarily indicate
wheter or not it unique, we cannot use the given name.

In order to meet the requirements, as well as having a *unique* CN not
only within a single federation, but across all federations, we concatenate
the given name (to meet SLIC/MICS reqs) as well as ePPN to meet our own
requirements.

Signed-off-by: Henrik Austad <henrik@austad.us>