list all available certificates to the user.
A while back, it was decided that listing all CSRs could be a security risk
as a user could be tricked into signing the wrong CSR. For that reason, the
list of CSRs was removed along with the list of signed certificates.

The latter proved to be not necessary, and for that reason, it has been
reintroduced to the system.

Signed-off-by: Henrik Austad <henrik@austad.us>