Remove dependency to IdPMap.php
The IdPMap adds the IdP-url as one of the attributes. However, this causes
ConfusaAttributeMap to depend not only on IdPMap, but on the order in which
the maps are called. This should be avoided at all costs as detailed
knowledge to the code is required in order to understand why the
ConfusaAttributeMap may fail to execute properly.

Signed-off-by: Henrik Austad <henrik@austad.us>