Removed verification of CSRs when using remote API
This is not needed any more because the Comodo API will overwrite all
subject fields and extensions in the issued certificate. We specifiy what
will be in the cert by the API call. Unfortunately, this changed the whole
sign_key() method. Took the opportunity to remove multiple return points.
Whitespaces are hard to handle, because the file contains a mixture of tabs
and whitespaces for indentation :(.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>