Made CertManager abstract, added listing and retrieval functions.
CertManager is now an abstract class that can be extended by subclasses
with specialized behaviour for signing keys, listing certificates,
retrieving them, etc. Most class members were removed since they were
mostly used only once in operations for which they were available at the
time of calling anyways (sign_key, etc.). Thus, CertManager can be
instantiated directly at the beginning of the program-flow in the
framework, where it can be kept for all parts that need the functionality.
The config-file needs to be consulted for 'standalone' or 'remote' only
once in the instantiation of cert-manager. The managed person is updated
on-the-go and many ugly if/else constructs in the rest of the code are
removed.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>