Add admin interface to Confusa
The admin interface is responsible for adding NRENs, adding subaccounts for
access to the remote-signing CA and assigning insitutions to the NRENs.
Additionally, it can be used for changing the state of the institutions
(subscribed, suspended, unsubscribed), so we can later filter that on the
application-side to determine whether an institution may use Confusa.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>