Add caching of remote certificate lists and downloaded certificates
Use the MySQL DB as a cache. Certificate lists are stored for the duration
of a session or until the user uploads new CSRs etc. and certificates are
cached for 30 minutes. The expiry of a certificate in cache is not checked
by the application ;-) - but the DB-cleaning script might have an interest
in getting rid of expired cached certs.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>