Make caching of certificate list less prone to SQL injection
The idea of the bulk processing was to make it fast. But how many
certificates will a user ever have? Having >40 here lets it still be
reasonably fast, so normally it won't be a problem that for each INSERT a
new connection has to be established. If it turns out to become a
performance bottleneck, it can still be changed.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>