menu: update menu
The menu was in sore need for an update after the refactor-work
began. This moves several of the entries around, and re-order some
more.

It will add 2 new elements to the menu for non-authN users, namely the
about (nren-branded/specific) and the help as this is pages a user would
normally want to look at first.

Signed-off-by: Henrik Austad <henrik.austad@uninett.no>