Add touch_person method do CertManager
This method allows to add a person to the cert_user table if one of the
following two conditions is met:
The person is not already there
The person is already there, but has a smaller expiry time than is
specified at insertion.

This is practical if the time that person names are stored in the cert_user
table should be limited to the expiry times of certificates, CSRs, etc.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>