Change format of timeout values in Config
Instead of a SQL time string, use an array in the form of [0] =>
time_amount [1] => time_unit, like e.g. array(7, 'DAY') for 7 days. It is
much more convenient to edit for the end user and along with the date_add()
function in mysql it can be used for arbitrary time-calculations, from
seconds up to years, which is a real pain with addtime(current_timestamp,
<sql-time-string>).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>