Remove SMS-layer from Confusa
SMS is not a safe technology, and users may not have the
mobile-attribute set. Furthermore, as the number can be changed by
anyone having the user's credentials, this second layer authN-option
does not add any extra security, it only adds added hazzle for the user.

This commit removes *all* traces of SMS and mobile from the system.

Signed-off-by: Henrik Austad <henrik@austad.us>