tools: move certManager creation
When downloading a certificate, the header needs to be rewritten. When
we pass the person to the super-ContentPage, output will be written
(when in debug-mode) before the headers are set so we can download the
certificate.

Therefore: use the person passed along and create the download-page. If
nothing is downloaded, it is 'safe' to write "anything" to the user.

Signed-off-by: Henrik Austad <henrik@austad.us>