confusa_auth: Add bypass
When debugging, it is often useful to have some debugging
functionality. This patch allows for a system to be set in
unauth-mode. This is of course somewhat dangerous, but is also means
that you do not need a fully configured SAML-interface when working on
Confusa.

This should lead to cleaner development, as it becomes easier to develop
on a non-production system.