Add NREN to attributes
When we do AuthZ we need both NREN and organization for an
admin. Currently, we have been using the organization as both, ignoring
the IdP entirey. However, it is wrong to use the IdP directly, as one
IdP can only belong to one NREN, but one NREN can have several IdPs.

This patch adds an initial mapping in ConfusaAttributeMap to connect the
IdPs to an NREN. confusa_auth sets this in person, and the about you
prints this out.

Note: This patch does *not* change any of the existing queries or tests,
it only adds more infrastructure.

Signed-off-by: Henrik Austad <henrik@austad.us>