Correctly handle the changed timeout format
The timeout format in confusa_config was changed to the interval-style,
i.e. a format like array(15, 'MINUTE'). The interval style has the
big advantage, that the represented timespan can be pretty arbitrary,
stretching from microseconds to years without any date/time distinction
and that it is better readable by humans than SQL-datetime
strings.

However, the code for insertion in cert_manager_standalone and for cleaning
up the CSRs in clean_db.sh were not adapted to the timeout-format. This is
changed with this bugfix.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Signed-off-by: Henrik Austad <henrik@austad.us>