download_certificate: add revoke-cert button when listing certificates
When listing the certificates, the user is given several options
(download, email, inspect and delete). It would be natural to add a link
to revocation here as well. As this means we have to mix POST and GET,
this can only be done via a FORM-button.

As a future tweak, revoke_certificate.php should accept a GET-url for
revoking single certificates, as this would mean passing an url, making
the per-certificate information consistent.

This latter point also means we can pass an url to a user with 'visit
this URL to revoke your certificate', if that is ever needed.

Signed-off-by: Henrik Austad <henrik@austad.us>