Add stylist interface to the NREN admin interface
The stylist interface can be used by the NREN admin to change the help and
about texts as well as the custom CSS that is to be used when displaying
Confusa for the constituency of the NREN.

The CSS is written to the file css/custom/<nren_name>/custom.css. This is
simpler than inserting it in the DB, because the CSS often contains
problematic characters such as {,#,;,.,* which would have to be checked
very carfully upon insertion into the DB. The file is never loaded into
executable space and only interpreted by the browser of the end-user as a
CSS, making it relatively unproblematic in file form.

Documentation and possibility to upload a custom logo are still to be
added to the stylist.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>