Delete previously stored custom logos from the NREN subfolder
This is not solved by an overwrite, because for instance there may be a
logo called custom.gif and the admin uploads a new logo, e.g. called
logo.png which will become custom.png. Now, in the folder there are
custom.gif and custom.png and only one of them will be displayed; dependant
on the file walking order the wrong one.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>