Cleanup CertManager_Online and add queries for certificate validity.
Filter out certificates in cert-listings if they are not valid and filter
out pending ones as well in common_name based queries. Once the validity
time of certificates is perfectly fixed, the datetime after which
returned certificates must not have been *ordered* can be passed to the
POST-API, for now we have to filter param-based.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>