person: change (get|set)_idp to (get|set)IdP
This change also includes a test for whether or not we're in
auth_bypass. Since the IdP is available from the session-array, we do
not need the setIdP when Confusa is coupled with SimpleSAML.

To retain a working setup when Confusa is placed in auth_bypass-mode,
the function remains, but will log errors when the IdP is set outside
this mode.

Signed-off-by: Henrik Austad <henrik@austad.us>