output: do not add spaces in help-link for different IP-addresses
When the current IP-address differs from the address where the CSR was
uploaded, Confusa will mark the address in red, bold letters and display
a [?] after it. This is so the user can 'ask' for why the address is
formatted in such a way.

By having spaces between '[' '?' and ']', the link tends to be split
between lines in the tables. Without the spaces, the link will be
grouped together.

Signed-off-by: Henrik Austad <henrik@austad.us>