Confusa frontpage: Updating text, adding FAQ like description
Confusa should give the user a reason to want to sign in with her identity
provider, since this is a rather sensitive action. Therefore Confusa should
explain what it does, how it treats the personal data of the user and why
it does what it does. Maybe the frontpage has to much text and the Q&A
sections should be made expandable (with a [+] on the left hand side for
instance).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Signed-off-by: Henrik Austad <henrik@austad.us>