Front page: Restore W3C-XHTML strict validity
The validator does not like the inline JavaScript, since < is only allowed
for tags. But there is a hack around that: wrap the JavaScript block in a
CDATA statement, since CDATA may contain arbitrary characters. Older
browsers may not understand the CDATA statement - that's why it is "hidden"
in a JavaScript comment.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>