key_(up|down)load - do not allow for usage in auth-mode.
When Confusa is in 'production mode', i.e. auth-bypass is not set, the
anonymous upload and download-pages should not be accessible. However,
when the system is placed in auth_bypass, this normally indicates that
we're looking at a test-system, and for those systems, having the option
of simple up/download is essential.

This patch adds a 'not-found' wrapper that rewrites the header and dumps
a subset of the standard 404-message to the user when the system is
*not* in 'auth_bypass'.

Signed-off-by: Henrik Austad <henrik@austad.us>