table_create: the auth-key of (csr|cert)_cache must be unique.
The auth-key of the certificate or CSR is actually the hash of the
public key, and should therefore be unique. The MICS profile state that
we *may* reissue a certificate, but this means that the certificate must
be gone from the database before the user is allowed to do so.

We need this because Confusa will panick when the auth_key is not unique.

Signed-off-by: Henrik Austad <henrik@austad.us>