sign_keys.sh: script for signing CSRs in standalone-mode
PHP is not great at handling openssl-bindings. So, in many cases we
*must* escape to shell, and in other cases we *think* we can avoid it,
but PHP then does such a crappy job, that we end up escaping to shell
anyway.

Since openssl_sign need to be given a serial for signing, and things do
not work well anyway, we do the whole signing step through shell_exec().

This script is the target of that. To prevent as many shell-code
injections as possible, it takes the auth_key and a temporary-file as
argument. It will then proceed to retrieve the CSR from the database,
signing the CSR and storing the result in the second parameter, a
temporary file created by the caller.

Signed-off-by: Henrik Austad <henrik@austad.us>