confusa_auth: Get logout-link from simplesamlphp location, don't hardcode
The logout link was hardcoded to the URL /simplesaml/saml2/sp/initSLO,
which means that simplesamlphp always has to be located at the URL. If
simplesamlphp is in path bla.com/simplesamlphp/ instead of
bla.com/simplesaml/ the previous logout link would point to a non-existing
page. This was easy to fix by just consulting the simplesamlphp configuration
where it actually is located.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>