maintenance-mode: allow confusa to be placed in maint-mode
When Confusa is placed in maintenance mode, it will not accept any
interaction with users except notifying that the site is under
maintenance and that it will be back up again 'shortly'.

This is useful when running scripts that require a consistent and static
database for the duration of the execution (e.g. db-cleanup, backup and
system upgrade).

Signed-off-by: Henrik Austad <henrik@austad.us>