Confusa_Auth: Remove deprecated authN files
OAuth_Auth will be redone according to the new authN class structure, but
currently I have no good method of testing it so this is going to be done
once OAuth based clients are in development again. Until then, the
subsystem's (simplesamlphp's) OAuth functionality may have improved also,
so no point in doing it now.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>