Root_cert: Show Terena/Comodo root cert and CRL when in online mode
If the user clicks a link to view them in the browser or download them,
the online-certificates are fetched using PHP curl, saved in tmp and
delivered to the user. Maybe this can be improved, for the moment it helped
not completely obfuscating the code. One could arguably also move the whole
functionality to CertManager, but then the root_cert page needs the content
of the files at one position, a URL to the files at another and the path to
the files at a third position and things would still be a bit torn apart.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>