I18N: Add pre-processing of template by Translator to content_page
Only if the subclassing content-page defined a dictionary for itself.
Otherwise the page will be delivered as usual, without predecoration by a
dictionary. This behaviour is unobtrusive and liberal - because it means we
don't have to have dictionaries for every page, but can add them
incrementally. The pages without dictionaries will just be delivered as
they were created - usually in something close to English.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>