I18N: Add available and default language fields to the config template
The default-language can be used for fallback if no suitable language is
found and the available-languages tell Confusa, which languages it can
expect to find in the dictionary files. Of course, this won't protect
against entries not being available in the dictionaries in the end.

But in case of a dictionary miss, the Translator will automatically try to
find the entry in the default language in the dictionary. If that entry is
also missing, the text with the looked up but missed tag will be missing
from the displayed page.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>