person: NREN admin cannot rely upon the entitlement attribute.
The problem is the initial bootstrapping when the first NREN admin wants
to create the map from local attributes to nren-specific attributes.

If this relies upon entitlement, the nren-admin can *never* acheive that
initial status, which we do not want. The NREN is supposed to have so
much sense that they can be trusted with this.

Signed-off-by: Henrik Austad <henrik@austad.us>