NREN-Admin: Use subscriber ID instead of name in edit/delete operations
It is also fairly safe to export it to the user, given that the checks if a
subscriber belongs to the NREN, of which the logged on user is the admin,
are still performed. Also, this gives the user a really unique identifier
with which she can refer to the subscriber, if she has to report something.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>