confusa_auth: decorate the person with attributes using the map
Use the retrieved map from the database to decorate the person. If the
map is not set and the user is an NREN-admin, provide a mechanism for
warning and continue.
If the user is a 'normal' user (i.e. a user not allowed to change the
map), throw a new MapNotFoundException.

Signed-off-by: Henrik Austad <henrik@austad.us>