framework: catch all unhandled exceptions from content-page process()
The process()-function should be self-contained wrt to exception
handling (i.e. no exception should ever propagate from process() and
into framework).

As this can be quite difficult to handle consistently save adding a
catch(Exception $e) (which is *not* the right way of doing this), a
protective try-catch is added around process() and we can therefore
notify the user that something is wrong without tearing down the entire
webface and dumping an ugly exception-rawlog.

Signed-off-by: Henrik Austad <henrik@austad.us>