create-database: Add possibility to drop the user in config at any rate
If the user is present, but the password for her in config has changed,
with the old method the user's real password and the password in
confusa_config.php will not match up any more. The easiest way to solve
this is to drop the user and add her anew.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>