CSS: Makeover the layout of the message containers
The containers are now free-floats in the page layout. Thus, they can
contain nested floats such as a left-floating icon. This was necessary to
keep the positioning of the different icons nice-looking and not having the whole
layout collapse if the a message container does not have any text (for
instance if Framework::error_output("") is called).

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>