Input: Replace \n in texts to <br /> for storage in database
It is a good idea to keep newlines in freely defined texts, since line
breaks are an important structuring element. It is not a good idea to store it
in the DB, however. Therefore, convert it to HTML <br /> after removal of the
other HTML tags and add a method to convert it back to \n for nice display e.g.
in textareas.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>