DevDocs: add the attribute-map layout.
The attribute-mapping allows an administrator to create a map, which in
turn will affect the users directly. The order of which the backend
processes this is important as a dedicated subscriber-map will overrule
a NREN-map.

This diagram tries to illustrate the flow of data, allthough the
subscriber vs. NREN priority is missing.

Signed-off-by: Henrik Austad <henrik@austad.us>