robot: add DTD to public location.
When we send and receive lists in the xml-format, we need to agree upon
some standard. This DTD must be publicly available so the remote sites
can do basic validation of the document.

Note, this DTD does not enforce all restrictions. For one, it does not
force valid date-formatting, nor does it assert that the element-count
is correct.

All the DTD is, is a syntactic guide to the elements needed in the file.

Signed-off-by: Henrik Austad <henrik@austad.us>