mdb2_wrapper: add better logging when query fails.
If the query should fail, we should log as much info as possible without
confusing the user.

This change adds more info to the logline as well as adding an
error-code to the message given the user so we can trace the problem in
the logs. If confusa is in debug-mode, additional info is logged (in
most cases it is a duplicate of the getUserInfo() content).

Signed-off-by: Henrik Austad <henrik@austad.us>