confusa_auth: add link to update-map page.
It is all well and good that the framework screams with rage if the map
is not set, but it should short circuit the path to where the map can be
created.

This updates the error-message from confusa_auth where an admin is
notified that the map is not set by including the url to handle_map.

Signed-off-by: Henrik Austad <henrik@austad.us>