confusa_auth: construct orgname based on dc-attributes.
In most cases (or all?) the eduPersonOrgDN will be used for
subscriber-name. This is normally encoded on the form:

	o=Hogwarts, dc=hsww, dc=wiz

and the resulting string should then be

	hogwarts.hsww.wiz

This change adapts to the old style (where the name is set directly) or
this approach and constructs a sensible string to use as subscriber
name.

It is not possible to adapt dynamically to (very) different
encoding-styles through creative mapping at the moment, and most likely
it will never be unless there is a dire need for it.

Signed-off-by: Henrik Austad <henrik@austad.us>