person: do not let a person be subscriber(sub)admin if no subscriber is set.
If no subscriber is added to the person, it follows not only by
definition, but also by the CP/CPS for the Terena PCS that this person
*cannot* and *should not* be an administrator.

The story is slightly different for an NREN-admin, however, the
NREN-admin should be associated with a subscriber as well. For
consistency, this test has been added to all 3 cases.

Signed-off-by: Henrik Austad <henrik@austad.us>