Cert-Manager: Add methods for signing certs using the browser keystore
If certificates are to be signed from the browser keystore, they are often
transported in a different format than PKCS#10. For instance, Firefox
generates CRMF requests, while storing the private key directly in its
keystore and "waiting" for a matching public key.
And the signed public keys need to be "installed" in the browser somehow,
usually using browser-specific JavaScript-API calls.

Therefore CertManager was enhanced with three methods
- signBrowserCSR (accept CSRs in typical browser formats and sign them)
- pollCertStatus (report to the browser, if the signing process is
finished)
- getCertDeploymentScript (return a script that can install the signed cert
in the browser's keystore, usually JavaScript)

CertManager_online implements the methods, but current browser
compatibility is limited to Firefox and CRMF/CMMF certificate formats. The
deployment script contains a JavaScript variable containing the signed cert
in CMMF format and a function call.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>