Upload_CSR: Add form for browser key generation
The form will kick of a JavaScript function that will use the browsers'
crypto mechanisms to create a certificate.

Then in a callback function the cert-request is passed to process_csr which
passes it to certmanager which passes it to the CA. The CA returns an
identifier ("order_number" or "auth_key"), back through certManager,
processCSR.php, upload_csr.tpl to the JavaScript status poller.

The JavaScript status poller does a meta-refresh like reload of the page
every ten seconds to check if the processing is already finished. If it is,
the form controls are made invisible and a link to install the certificate
in the browser is presented instead.

When the user clicks the link, again the "order_number" is propagated
through process_csr.php to the CertManager which returns a so called
deployment-script, a directly in the browser executable script-string,
usually in JavaScript that will contain the certificate as a variable and
install it in the browser's keystore. The script-string is pushed to the
template, where it will lead to execution.

Currently all of the above works exclusively for Firefox and exclusively in
Online mode.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>