cert_manager: verify attributes for person
When we are about to sign a certificate, we must assert all required
attributes. This will now be handled by CertManager, and a
KeySignException will be thrown if the attributes are not properly set.

This test was previously locaed in process_csr.php in the www/
directory. However, this test is crucial for signing. CertManager should
not rely upon the caller to do this test prior to invoking sign_key().

Signed-off-by: Henrik Austad <henrik@austad.us>