admin: Add more rigid tests to deleteAdmin()
When deleting an admin, we must make sure that the admin is part of the
correct NREN (as we only require the naming to be unique within a NREN
and mistakes *will* be made when adding new admins).

If the admin is part of the correct NREN, we must also verify that the
admin requesting the delete, has the rights to do so. This is currently
handled in the pre_process() section, but this test is critical and
*must* be done by the delete-function, otherwise we are going to miss
this test when we expand the admin-functionality later on.

Signed-off-by: Henrik Austad <henrik@austad.us>