person: remove unnecessary test from getAdminStatus()
By testing to see if 'admin' in the returned row matches the ePPN of
person is redundant as we fetch rows based, among other things, on the
ePPN. So *every* row we test for there, *will* have ePPN set. This test
will therefore always evaluate to true, and is therefore redundant.

Signed-off-by: Henrik Austad <henrik@austad.us>