robot: get admin, subscriber and nren ID from database
This uses a somewhat ugly join to produce a table with the 3 desired
id's. It can probably be done differently, but it works, and it is not
too slow. Anyhow, this is not an action that we will perform very often,
and the set of rows to join will be bounded, forcing the time taken to
be low-ish.

Signed-off-by: Henrik Austad <henrik@austad.us>