person: move subscriber-test to appropriate section.
When testing for admin-status, the low-level getAdminStatus() assumed
that all administrators must have the subscriber-name set. This is wrong
as an NREN-admin must create the mapping, and until the mapping has been
created, it is not possible to retrieve this name.

Furthermore, an NREN-admin does not belong to any subscriber per. se, so
adding this restriction to NREN-admins is also wrong from a conceptual
point of view.

Signed-off-by: Henrik Austad <henrik@austad.us>