Root_cert: Do not use filenames from standalone-config for online download
When downloading, the Comodo root-certs were cached in /tmp/$ca_cert_name,
where $ca_cert_name was the path from confusa_config. However that value
may not and need not be configured when running Confusa in Online mode.

Therefore use a fixed value for caching the certificates in /tmp/ and for
the filenames of the certificates when they are downloaded.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>