Rename bootstrap_admin.php to bootstrap_nren.php
What we actually bootstrap is a NREN, not an admin. Thus, the old name
was very confusing, because the script is supposed to be called only once
per NREN with an initial admin. Afterwards, the script will report an
error, stating that the NREN is already present.

The old name may have lead administrators to believe that they could just
arbitrarily add new administrators with the script.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>