robot: when no subscriber-admin is found, detect known issues and report.
The rather ugly query for finding the id's for all parties (nren,
subscriber and admin) will fail if the subscriber-admin has no
subscriber set. This is probably due to a db-inconsistency (the result
of manually adding a subscriber-admin directly into the database).

This change will detect this error and provide feedback to both user and
the log so the operational team can resolve it.

Signed-off-by: Henrik Austad <henrik@austad.us>