admin_db: add name and email to admins-table
When we deal with admins, we should store the complete name and email as
this is useful for user-feedback and logging.

This change adds this by adding the values the first time the admin logs
into confusa. The values are needed by the robotic-interface, and it is
safe to do so, as the admin *must* log in at least once in order to
upload a new certificate. Thus, when the RI needs the values, they will
be present.

Signed-off-by: Henrik Austad <henrik@austad.us>