Revoke-Certificate: Do not consume the auth_keys from POST parameters
This is dangerous. Since the POST parameters can include arbitrary data and
can be changed very easily, somebody could just try out to send different
order_numbers or auth_keys. The certificates belonging to these
order_numbers or auth_keys would be revoked, no matter to whom they belong.
We don't want that.

Instead, the following is done now: Upon fetching the certificate list, a
mapping common_name <-> auth_key is stored in the session. If the user
chooses to revoke certificates of an end entity, only the common-name of
the end-entity is POSTed to Confusa. Now, the corresponding order-numbers
can be looked up in the session and sent e.g. to Comodo for revocation. If
the common_name is tampered with, no auth_key will be found in the session,
so this will only lead to an error or a no-op.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>