robot: main robot interface
The following acceptance-tests are done:
- The connection is via SSL on SSLv3 or TLSv1 only
- The certificate exists in the database
- only one certificate

From this, an admin-person is created which is used to create a
CertManager.

Then, based on input, a list of all valid certificates is returned, or
the certificates for the listed users (belonging to the current
subscriber) is revoked.

In both cases, a list of ePPN's along with a counter is returned either
indicating the number of certificates revoked, or the number of active
certificates.

Signed-off-by: Henrik Austad <henrik@austad.us>