RI: return result even if the user was not found
Return should be the exact same list as provided by the client, and
count should be the number of revoked certificates. If this number is 0,
then it means that the user had no valid certificates (or the user was
not found, which pretty much means the same thing from Confusa's point
of view).

Signed-off-by: Henrik Austad <henrik@austad.us>