Revocation: Restructure the templates and the PHP
Do not show any revocation options when arriving with a GET request (that
will be either automatic revocation or from the download page), just a
success message or an error message depending on if it worked.

Otherwise show one of NREN/subscriber perspectives if in admin mode. If not
in admin mode, just offer the user to revoke her own certificates.

This should make seeing what's going on more transparent.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>