robot_www: do not test for subscriber-admin capabilities
When we have a valid robot-certificate, we have to trust the
admin-interface in confusa to verify the user properly. It does not make
sense to add this attribute to the person at this stage as the *only*
persons allowed in here is subscriber-admins. So, by being allowed to
create a channel, the user is by definition, a subscriber-admin.

Signed-off-by: Henrik Austad <henrik@austad.us>