confusa_auth: Add extra feedback to the user when missing the email.
The email is required when  doing both revocation and signing, thus both
of the  "core" operations in Confusa.  When the email  is not available,
the user should be given warning before he enters the signing-area, only
to be told that it won't work.

This change adds an extra test to decoratePerson, so the warning will
trigger every time Confusa handles a session.

Signed-off-by: Henrik Austad <henrik@austad.us>