Custom-CSS/Logo: Use absolute path in confusa_config(_template)
This will partly fix the Tilburg requirement that the webserver must not
write to /usr/share/, but to /var/lib/ instead. Some larger changes
were necessary to enable it, because before we were just linking to the
respective files in www/.

Now the small helper scripts get_css.php and view_logo.php fetch the logo
and the CSS from their position on the server's harddrive, set a
corresponding header and echo the content. Instead of to the image
directly, framework links to view_logo.php?nren=<nren>. That sounds
complicated, but in reality Framework got much lighter by losing the
logo/CSS discovery functions and the small retrieval scripts are mostly
self-explanatory.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>