Standalone: Move all files writable by webserver to /var/lib/confusa/
This tosses around standalone signing and revocation quite a bit. The
concept of providing CRL and root-cert in www/ca has been thrown away
altogether. Instead the files are just saved in one place
(/var/lib/confusa/cert/) and retrieved from there, if the user wants to see
either the CRL or the root-cert.

The CRL filename is not configurable any more, but a confusa constant now.
It makes sense to configure the cert- and privatekey name and location,
because these are files that the user provides. But the CRL is generated
and written to by Confusa, so it's Confusa's good right to decide about the
name :)

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>