nren_admin: split the nren-admin view and add more fields for add subscriber
When we add a new subscriber, we add a lot of new fields into the
database. This change adds the interface for this, with a lot of
comments to the admin to aid in the completion.

This change splits the old view into 2 section
1) List the available subscribers. This is the default view.
2) Add new subscriber. The old section is completely rewritten with a
   lot of new fields added. A lot of help has been added to aid the
   admin.

Currently, the new information is not validated, nor added to the
database (coming soon).

Signed-off-by: Henrik Austad <henrik@austad.us>