Person: Preserve UTF-8 chars of the name attribute, map to ASCII for DN
To do so, do not pass the name attribute through htmlentities upon
decoration, but only when retrieving the values with getXXX(). Person
should store the attributes as close to what was received as possible and only
change them for the situation in which they are used.

Hence store them in UTF-8 format and map them to ASCII upon constructing
the DN. Thus, we also achieve a well-defined mapping to ASCII - instead of
just dropping characters, we can transfer them to their ASCII counterpart.

Furthermore, if we ever issue personal (e-mail) certificates, we can use
UTF-8 DNs.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>