nren_admin: add new subscriber the 'new' way.
This updates the addSubscriber() to handle all the new fields. Some
rudimentary checks are perfomed on the attributes. If the attribute is
deemed critical, an error is generated, otherwise it is set to a default
value (empty string).

If Confusa is in "grid-mode", the length is checked. At the moment, no
7-bit asci-test is performed.

The DB-queries are updated and the error-logging/output is improved
slightly.

Signed-off-by: Henrik Austad <henrik@austad.us>