CertManager_Online: Give more specific exception, if account info missing
The old behaviour was broken, because Framework would interpret the message
that the credentials are missing as a ConfusaGenException and set
renderError to true. So there was no way to specify the credentials
(except directly inserting them into the DB), which was rather bad.

With a specifc CGE_RemoteCredentialException that issue is solved and a
more specific error message can be displayed since the framework now knows
exactly what's the problem.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>