NREN-Admin: Do not sanitize DB-name and DN-name
If sanitizing, one would have to consider that a DB-name can look like
dc=uninett,dc=no or in any way. It really IS hard to predict, how the IdPs
export the eduPersonOrgDN and it is also hard to predict, how the DN-name
that admins want to have in the certificate subject will look like.

We should only be careful to pass everything through htmlentities when
displaying it or using only prepared statements when calling to the DB.

Signed-off-by: Thomas Zangerl <tzangerl@pdc.kth.se>