table_create: allow the contact-email to be longer than 30 characters.
Normally, 30 characters will be more than enough. However, we should be
generous with this field - we are not going to have an infinite number
of NRENs, using a longer field to allow for longer addresses will not be
a space-issue.

Furthermore, the contact *must* be set, we will need a contact point
for every NREN.

Signed-off-by: Henrik Austad <henrik@austad.us>